span: rip ranged integers out of `Span`
OMG! DONE!

While not as much as I'd hoped, there is a small but measurable
improvement to compile times for Biff. Before:

```
$ hyperfine -w1 --max-runs 3 --prepare 'cargo clean' 'cargo b' 'cargo b -r' ; A bart
Benchmark 1: cargo b
  Time (mean ± σ):      4.753 s ±  0.018 s    [User: 15.921 s, System: 2.204 s]
  Range (min … max):    4.733 s …  4.767 s    3 runs

Benchmark 2: cargo b -r
  Time (mean ± σ):      7.525 s ±  0.145 s    [User: 64.197 s, System: 2.727 s]
  Range (min … max):    7.372 s …  7.661 s    3 runs
```

After:

```
$ hyperfine -w1 --max-runs 3 --prepare 'cargo clean' 'cargo b' 'cargo b -r' ; A bart
Benchmark 1: cargo b
  Time (mean ± σ):      4.406 s ±  0.044 s    [User: 15.453 s, System: 2.218 s]
  Range (min … max):    4.360 s …  4.447 s    3 runs

Benchmark 2: cargo b -r
  Time (mean ± σ):      7.400 s ±  0.041 s    [User: 65.723 s, System: 2.705 s]
  Range (min … max):    7.358 s …  7.441 s    3 runs
```

Interestingly, total LLVM lines does go down for Biff from 719,419 to
716,473. I wish I could say that that suggests LLVM wasn't completely
seeing through the abstraction. Unfortunately, ripping out ranged
integers meant that I switched some code to using `SignedDuration`
and even `i64` (from `i128`) in some cases. So the comparison isn't
completely apples-to-apples. Still it's a nice reduction.

+++ src/fmt/util.rs
@@ -797,12 +797,12 @@ fn fractional_time_to_span(
     fraction: i32,
     mut span: Span,
 ) -> Result<Span, Error> {
-    const MAX_HOURS: i64 = b::SpanHours::MAX;
+    const MAX_HOURS: i64 = b::SpanHours::MAX as i64;
     const MAX_MINS: i64 = b::SpanMinutes::MAX;
     const MAX_SECS: i64 = b::SpanSeconds::MAX;
     const MAX_MILLIS: i128 = b::SpanMilliseconds::MAX as i128;
     const MAX_MICROS: i128 = b::SpanMicroseconds::MAX as i128;
-    const MIN_HOURS: i64 = b::SpanHours::MIN;
+    const MIN_HOURS: i64 = b::SpanHours::MIN as i64;
     const MIN_MINS: i64 = b::SpanMinutes::MIN;
     const MIN_SECS: i64 = b::SpanSeconds::MIN;
     const MIN_MILLIS: i128 = b::SpanMilliseconds::MIN as i128;
